前端完善

+++ src/main/resources/templates/userManage.ftl
@@ -105,12 +105,12 @@
                 </div>
                 <div class="modal-body">
                     <div class="input-group">
-                        <span class="input-group-addon">姓名</span>
-                        <input class="form-control" type="text" name="name" id="name">
+                        <span class="input-group-addon">姓名<span style="color:red">*</span></span>
+                        <input class="form-control" type="text" name="name" id="name" mustwrite="true" >
                     </div>
                     <div class="input-group">
-                        <span class="input-group-addon">密码</span>
-                        <input class="form-control" type="password" name="password" id="password">
+                        <span class="input-group-addon">密码<span style="color:red">*</span></span>
+                        <input class="form-control" type="password" name="password" id="password" mustwrite="true" >
                     </div>
                     <div class="input-group">
                         <span class="input-group-addon">部门</span>
@@ -177,6 +177,10 @@
             });
 
             $modal.find('.submit').click(function () {
+                if(!check()){
+                    alert("请填写所有必填项！");
+                    return;
+                }
                 var row = {};
                 $modal.find('input[name]').each(function () {
                     row[$(this).attr('name')] = $(this).val();
@@ -210,6 +214,19 @@
                 });
             });
         });
+
+        //批量校验input必填项
+        function check() {
+            var item=$modal.find('input[mustwrite="true"]');
+            // var item = $("input[mustwrite='true']", document.forms[0]);
+            for (var i = 0; i < item.length; i++) {
+                if ($.trim(item[i].value) == "") {
+                    item[i].focus();    //光标定位到未填input框中
+                    return false;
+                }
+            }
+            return true;
+        }
 
         function queryParams(params) {
             var user = { //这里的键的名字和控制器的变量名必须一直，这边改动，控制器也需要改成一样的
@@ -245,6 +262,10 @@
             },
             'click .remove': function (e, value, row) {
                 if (confirm('确定删除此用户 ？')) {
+                    if(row.role==true){
+                        showAlert('禁止删除管理员!', 'danger');
+                        return;
+                    }
                     $.ajax({
                         url: "/admin/del?id=" + row.id,
                         type: 'delete',
